refactor: use the same naming convention across guards
